Update index.html
corrections to syntax on meta tags. Meta tags checked again for expedience, as not showing on Google but did show on Google the first time around before I added the URL image. I will leave it for a few days for Google to catch up . I am hoping to get an image on my URL on the internet. I can always go back to the old meta title tags, but have changed it so a user searching for 'penalties' will find the site.